+++ resources/lang/ca/badges.php
@@ -16,4 +16,10 @@ return [
     'hp' => 'Punts de vida',
     'experience' => 'Experiència',
     'gold' => 'Or',
+    
+    'badge_exemplar' => 'Comportament excemplar',
+    'badge_exemplar_info' => 'Sense interaccions negatives! Aquesta costa de mantenir ... Ànims!',
+    'academic' => 'Reconeixement',
+    'academic_info' => 'Totes les notes per sobre de ',
+    'challenge_info' => 'Has superat tots els reptes de ":name", continua així!',
 ];