+++ resources/lang/es/badges.php
@@ -13,7 +13,14 @@ return [
     'attributes' => 'Atributos',
     'attributes_info' => 'valores positivos para buenas insignias y negativas para malas',
 
-    'hp' => 'PPuntos de vida',
+    'hp' => 'Puntos de vida',
     'experience' => 'Experiencia',
     'gold' => 'Oro',
+
+    'badge_exemplar' => 'Alumnado ejemplar',
+    'badge_exemplar_info' => 'Sin interacciones negativas. Esta cuesta de mantener... ¡Ánimos!',
+    'academic' => 'Reconocimiento',
+    'academic_info' => '¡Todas las notas per encima de ',
+    'challenge_info' => 'Has superado todos los desafíos de ":name", ¡continúa así!',
+
 ];